final_project/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
matplotlib==3.7.2
python-dotenv==1.0.0